refactor: move NetworkNotePricer into miden-tx
miden-tx depends on both miden-standards and miden-agglayer, so the pricer
(with NotePricingError and CostLookupFn) lives there now and its lookup
defaults to the widest one, AgglayerNote::note_cost - a single pricer covers
agglayer and standard notes out of the box, per the review discussion. The
standards crate keeps the cost data: NoteConsumptionCost, NoteCost, the
tables and StandardNote::note_cost.

The pricer tests move along, and the agglayer claim-price test folds into
them (exercising the new default lookup), which also removes the fee()
test-helper duplication across crates.
